Move App and related components to separate folder
