+++ src/App/Components/AppNavigation/index.ts
@@ -1,0 +1,1 @@
+export { AppNavigation } from "./AppNavigation";
